Initial layer-shell support
Requires `layer-shell` feature to support, and patched GTK from
https://github.com/pop-os/gtk4/tree/layer-shell_impish.

This is generally working, but some things could be improved. Some of
the methods implemented for `LayerShellWindow` lack logic that exists in
`GtkWindow` that may be important. (For instance, some things related to
CSS). And some things may require more private functions from GTK.

The hard part is getting the necessary work upstreamed in some form.