hack: a temprorary fix for power applet sizing
The power applet currently doen't load properly in the panel.
The icon doesn't load and the allocated space is far too big.

This seems to be caused by the power applet only rendering once,
all of the other applets render more than once to fully load, all
this does is add a dummy command at app start to casuse a redraw.

My guess is that this bug is caused somewhere in iced_sctk, but
that is just a guess. As iced_sctk is currelty being moved as I
write this message, this is a temporary fix until I can look into
the issue further (it might not be iced_sctk at all).