fix: crash when no icon found
Specific case when using vscode, the open file/folder dialog
doesn't succeed in looking up the icon on my system. This would
cause lookup to return None, thus the ret vector is empty.
Since we removed the id before the lookup, the append below would
append an empty vec. This would cause an error in AddTopLevel
message handler when trying to remove the first item of the
returned vec.

Move the id remove directly within the map, so it only gets
removed if we succeed in looking up the icon.