feat(panel-button): add applet config
This PR adds an optional environment variable to the cosmic-panel-button that allows us to specify how a panel button should present itself (either as 'icon' or as 'text').

The PR also adds the same tracing log that the other applets have, with an initial message notifying that the applet has started.