Update volume sliders when changing volume with media keys
Recently a change was made to update the audio applet icons when the
volume changes by an action from outside the applet, but the volume
sliders still didn't update.

So for example, if you have the audio applet open, and use media keys on
your keyboard to change the volume, the volume sliders will now update
in the applet to reflect the change.