[WIP] status-area: Handle menuless items properly
SNI protocol does not mandate the presence of a DBus menu.
The `ItemIsMenu` method returns whether the item is a menu or not.
But we should also handle items that don't provide a menu, even if they
don't advertise this fact through `ItemIsMenu`.

If the item isn't a menu (or doesn't have one), then:
- Left click should call the `Activate` method on the item.
- Right click should call the `ContextMenu` method on the item.

This is a work-in-progress implementation to make the applet
respect the above.