time: Tick per sec if strftime has seconds spec
It's possible to have a discrepancy between strftime and the
show_seconds setting. If show_seconds is disabled but a user's strftime
formatter has a seconds specifier, then the time applet won't update
seconds because it's internally ticking per minute.

The fix is to ensure that the time applet updates per second regardless
of the user's setting if strftime has a seconds specifier. This is an
internal flag and shouldn't affect the user's show_seconds setting.